Added new migration Files and exported tables to SQL using Codefirst
