+++ SneakerDrop.Code/SneakerDropMigrate/20190131045948_second_migration.cs
@@ -1,0 +1,249 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace SneakerDrop.Code.SneakerDropMigrate
+{
+    public partial class second_migration : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.EnsureSchema(
+                name: "User");
+
+            migrationBuilder.EnsureSchema(
+                name: "Store");
+
+            migrationBuilder.CreateTable(
+                name: "Address",
+                schema: "User",
+                columns: table => new
+                {
+                    AddressId = table.Column<int>(nullable: false)
+                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                    Street = table.Column<string>(maxLength: 50, nullable: false),
+                    City = table.Column<string>(maxLength: 50, nullable: false),
+                    State = table.Column<string>(maxLength: 50, nullable: false),
+                    PostalCode = table.Column<string>(maxLength: 50, nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Address", x => x.AddressId);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "Orders",
+                schema: "User",
+                columns: table => new
+                {
+                    OrderId = table.Column<int>(nullable: false)
+                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                    OrderGroupNumber = table.Column<int>(nullable: false),
+                    Quantity = table.Column<int>(nullable: false),
+                    ShippingStatus = table.Column<string>(maxLength: 50, nullable: false),
+                    Timestamp = table.Column<DateTime>(nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Orders", x => x.OrderId);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "Listing",
+                schema: "Store",
+                columns: table => new
+                {
+                    ListingId = table.Column<int>(nullable: false)
+                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                    UserSetPrice = table.Column<decimal>(nullable: false),
+                    Quantity = table.Column<int>(nullable: false),
+                    Size = table.Column<string>(maxLength: 50, nullable: false),
+                    OrdersOrderId = table.Column<int>(nullable: true)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Listing", x => x.ListingId);
+                    table.ForeignKey(
+                        name: "FK_Listing_Orders_OrdersOrderId",
+                        column: x => x.OrdersOrderId,
+                        principalSchema: "User",
+                        principalTable: "Orders",
+                        principalColumn: "OrderId",
+                        onDelete: ReferentialAction.Restrict);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "Payment",
+                schema: "User",
+                columns: table => new
+                {
+                    PaymentId = table.Column<int>(nullable: false)
+                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                    CCNumber = table.Column<long>(nullable: false),
+                    CCUserName = table.Column<string>(maxLength: 50, nullable: false),
+                    Month = table.Column<int>(nullable: false),
+                    Year = table.Column<int>(nullable: false),
+                    CVV = table.Column<int>(nullable: false),
+                    OrdersOrderId = table.Column<int>(nullable: true)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Payment", x => x.PaymentId);
+                    table.ForeignKey(
+                        name: "FK_Payment_Orders_OrdersOrderId",
+                        column: x => x.OrdersOrderId,
+                        principalSchema: "User",
+                        principalTable: "Orders",
+                        principalColumn: "OrderId",
+                        onDelete: ReferentialAction.Restrict);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "Product",
+                schema: "Store",
+                columns: table => new
+                {
+                    ProductInfoId = table.Column<int>(nullable: false)
+                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                    ProductTitle = table.Column<string>(maxLength: 50, nullable: false),
+                    Brand = table.Column<string>(maxLength: 50, nullable: false),
+                    Type = table.Column<string>(maxLength: 50, nullable: false),
+                    Description = table.Column<string>(maxLength: 500, nullable: false),
+                    DisplayPrice = table.Column<int>(nullable: false),
+                    ReleaseDate = table.Column<string>(maxLength: 50, nullable: false),
+                    Color = table.Column<string>(maxLength: 50, nullable: false),
+                    ListingId = table.Column<int>(nullable: true)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Product", x => x.ProductInfoId);
+                    table.ForeignKey(
+                        name: "FK_Product_Listing_ListingId",
+                        column: x => x.ListingId,
+                        principalSchema: "Store",
+                        principalTable: "Listing",
+                        principalColumn: "ListingId",
+                        onDelete: ReferentialAction.Restrict);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "User",
+                schema: "User",
+                columns: table => new
+                {
+                    UserId = table.Column<int>(nullable: false)
+                        .Annotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn),
+                    Firstname = table.Column<string>(maxLength: 50, nullable: false),
+                    Lastname = table.Column<string>(maxLength: 50, nullable: false),
+                    Username = table.Column<string>(maxLength: 50, nullable: false),
+                    Email = table.Column<string>(maxLength: 50, nullable: false),
+                    Password = table.Column<string>(maxLength: 50, nullable: false),
+                    AddressId = table.Column<int>(nullable: true),
+                    OrdersOrderId = table.Column<int>(nullable: true),
+                    PaymentId = table.Column<int>(nullable: true),
+                    ListingId = table.Column<int>(nullable: true)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_User", x => x.UserId);
+                    table.ForeignKey(
+                        name: "FK_User_Address_AddressId",
+                        column: x => x.AddressId,
+                        principalSchema: "User",
+                        principalTable: "Address",
+                        principalColumn: "AddressId",
+                        onDelete: ReferentialAction.Restrict);
+                    table.ForeignKey(
+                        name: "FK_User_Listing_ListingId",
+                        column: x => x.ListingId,
+                        principalSchema: "Store",
+                        principalTable: "Listing",
+                        principalColumn: "ListingId",
+                        onDelete: ReferentialAction.Restrict);
+                    table.ForeignKey(
+                        name: "FK_User_Orders_OrdersOrderId",
+                        column: x => x.OrdersOrderId,
+                        principalSchema: "User",
+                        principalTable: "Orders",
+                        principalColumn: "OrderId",
+                        onDelete: ReferentialAction.Restrict);
+                    table.ForeignKey(
+                        name: "FK_User_Payment_PaymentId",
+                        column: x => x.PaymentId,
+                        principalSchema: "User",
+                        principalTable: "Payment",
+                        principalColumn: "PaymentId",
+                        onDelete: ReferentialAction.Restrict);
+                });
+
+            migrationBuilder.CreateIndex(
+                name: "IX_Listing_OrdersOrderId",
+                schema: "Store",
+                table: "Listing",
+                column: "OrdersOrderId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_Product_ListingId",
+                schema: "Store",
+                table: "Product",
+                column: "ListingId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_Payment_OrdersOrderId",
+                schema: "User",
+                table: "Payment",
+                column: "OrdersOrderId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_User_AddressId",
+                schema: "User",
+                table: "User",
+                column: "AddressId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_User_ListingId",
+                schema: "User",
+                table: "User",
+                column: "ListingId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_User_OrdersOrderId",
+                schema: "User",
+                table: "User",
+                column: "OrdersOrderId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_User_PaymentId",
+                schema: "User",
+                table: "User",
+                column: "PaymentId");
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "Product",
+                schema: "Store");
+
+            migrationBuilder.DropTable(
+                name: "User",
+                schema: "User");
+
+            migrationBuilder.DropTable(
+                name: "Address",
+                schema: "User");
+
+            migrationBuilder.DropTable(
+                name: "Listing",
+                schema: "Store");
+
+            migrationBuilder.DropTable(
+                name: "Payment",
+                schema: "User");
+
+            migrationBuilder.DropTable(
+                name: "Orders",
+                schema: "User");
+        }
+    }
+}
